Add organization-wide CODE_OF_CONDUCT.md file
We already have an older version of the Contributor Covenant in some of our repos. I'm upgrading us to the latest version and adding it to this .github repo so that it applies to all our repos. See Slack thread: https://hypothes-is.slack.com/archives/C4K6M7P5E/p1658850958942539